Rename Modules and packages accross the repository to follow PEP 8 naming convention
