+++ examples/Complubot/beep_test.py
@@ -1,5 +1,5 @@
 from time import sleep
-import HALduino.halduino as halduino
+import halduino.halduino as halduino
 
 
 def loop():
