+++ examples/Complubot/hit_and_rotate_test.py
@@ -1,23 +1,24 @@
 from time import sleep
 
-import HALduino.halduino as halduino
+import halduino.halduino as halduino
 
 
 def set_engine(direction: int):
     if direction == 0:
         halduino.setSpeedEngines(0, 0)
         sleep(500)
-        halduino.playBuzzer(330, 1000)
         halduino.setSpeedEngines(-100, -100)
         sleep(500)
         halduino.setSpeedEngines(100, 0)
-        sleep(1200)
+        sleep(100)
+        print('Rotate!')
     elif direction == 1:
         halduino.setSpeedEngines(100, 100)
+        print('Forward')
 
 
 def loop():
-    if halduino.getLightSensor() < 100:
+    if halduino.getUS() < 30:
         set_engine(0)
     else:
         set_engine(1)
