+++ examples/MBot/hit_and_rotate_test.py
@@ -1,10 +1,17 @@
-import HALduino.halduino as halduino
+from time import sleep
+
+import halduino.halduino as halduino
 
 
 def set_engine(direction: int):
     if direction == 0:
         halduino.setSpeedEngines(0, 0)
-        print('STOP!')
+        sleep(500)
+        halduino.setSpeedEngines(-100, -100)
+        sleep(500)
+        halduino.setSpeedEngines(100, 0)
+        sleep(500)
+        print('Rotate!')
     elif direction == 1:
         halduino.setSpeedEngines(100, 100)
         print('Forward')
