+++ examples/MBot/led_test.py
@@ -1,6 +1,6 @@
 from time import sleep
 
-import HALduino.halduino as halduino
+import halduino.halduino as halduino
 
 
 def loop():
