+++ examples/MBot/line_follow_test.py
@@ -1,4 +1,4 @@
-import HALduino.halduino as halduino
+import halduino.halduino as halduino
 
 
 def loop():
