+++ examples/MBot/stop_and_go_test.py
@@ -1,0 +1,17 @@
+import halduino.halduino as halduino
+
+
+def set_engine(direction: int):
+    if direction == 0:
+        halduino.setSpeedEngines(0, 0)
+        print('STOP!')
+    elif direction == 1:
+        halduino.setSpeedEngines(100, 100)
+        print('Forward')
+
+
+def loop():
+    if halduino.getUS() < 10:
+        set_engine(0)
+    else:
+        set_engine(1)
