+++ examples/SergioRobot/stop_and_go_test.py
@@ -1,4 +1,4 @@
-import HALduino.halduino as halduino
+import halduino.halduino as halduino
 
 
 def set_engine(direction: int):
