+++ translator/strings/translator_strings.py
